Split TLS into several files
